+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * [#5057](https://github.com/bbatsov/rubocop/pull/5057): Add new `Gemspec/RequiredRubyVersion` cop. ([@koic][])
 * [#5087](https://github.com/bbatsov/rubocop/pull/5087): Add new `Gemspec/RedundantAssignment` cop. ([@koic][])
 * Add `unannotated` option to `Style/FormatStringToken` cop. ([@drenmi][])
+* [#5077](https://github.com/bbatsov/rubocop/pull/5077): Add new `Rails/CreateTableWithTimestamps` cop. ([@wata727][])
 * Add new `Style/ColonMethodDefinition` cop. ([@rrosenblum][])
 
 ### Bug fixes
@@ -3048,5 +3049,6 @@
 [@sambostock]: https://github.com/sambostock
 [@asherkach]: https://github.com/asherkach
 [@tiagotex]: https://github.com/tiagotex
+[@wata727]: https://github.com/wata727
 [@marcandre]: https://github.com/marcandre
 [@walf443]: https://github.com/walf443